voxel_to_mesh_pipeline/requirements.txt
# Core deep learning
torch>=2.0.1
torchvision>=0.15.2

# Scientific computing
numpy>=1.24.3
scipy>=1.10.1
scikit-image>=0.20.0

# Data processing
pandas>=2.0.3
h5py>=3.9.0
pyyaml>=6.0

# Visualization and I/O
matplotlib>=3.7.1
imageio>=2.31.1
pillow>=9.5.0

# 3D processing
trimesh>=3.22.4
plyfile>=0.7.4

# Voxel processing
binvox-rw>=1.0.1

# Development and utilities
cython>=0.29.36
pytest>=7.4.0
tqdm>=4.65.0
setuptools>=68.0.0
wheel>=0.40.0

# Monitoring
tensorboard>=2.13.0

# Optional but recommended
# pyembree>=0.1.6  # For fast ray tracing (may have installation issues)
# open3d>=0.17.0   # Alternative 3D processing library
